+++ inst/examples/gr_kable_tests.R
@@ -7,8 +7,8 @@ if (require("kableExtra")) {
   # separate
   sep = gr_separate(spas, params = gr_get_params(reg = 'center'))
   
-  # summarize
-  vars = gr_summarize(sep)
+  # summarize from 1965 to 1990
+  vars = gr_summarize(sep, 1965, 1990)
   
   # test all variables
   tests = gr_test_vars(vars)